fix: 유저 요금 충전 / 조회 API 구현

+++ src/main/java/com/hhplus/concert/core/interfaces/v1/user/res/SelectUserAmountRes.java
@@ -4,6 +4,6 @@ import io.swagger.v3.oas.annotations.media.Schema;
 
 public record SelectUserAmountRes(
         @Schema(description = "남은 금액", defaultValue = "50000")
-        int amount
+        Long amount
 ) {
 }
