+++ src/main/java/com/hhplus/concert/core/interfaces/v1/user/res/UserAmountChargeRes.java
@@ -4,6 +4,6 @@ import io.swagger.v3.oas.annotations.media.Schema;
 
 public record UserAmountChargeRes(
         @Schema(description = "충전 완료 후 남은 금액", defaultValue = "50000")
-        int amount
+        Long amount
 ) {
 }
